// packages/interceptor/src/test.enum.ts
export enum CallLineEnum {
    n000001 = "000001",
    n000002 = "000002",
    n000003 = "000003",
    n000004 = "000004",
    n000005 = "000005",
    n000006 = "000006",
    n000007 = "000007",
    n000008 = "000008",
    n000009 = "000009",
    n000010 = "000010",
    n000011 = "000011",
    n000012 = "000012",
    // n000013 = "000013",
    // n000014 = "000014",
    n000015 = "000015",
    n000016 = "000016",
    n000017 = "000017",
    n000018 = "000018",
    n000019 = "000019",
    n000020 = "000020",
    n000021 = "000021",
    n000022 = "000022"
}